added pkg, and changed ips to local ip
