+++ FrontEnd/babel.config.js
@@ -1,7 +1,7 @@
 module.exports = function (api) {
   api.cache(true);
   return {
-    presets: ["babel-preset-expo", "module:metro-react-native-babel-preset"],
+    presets: ["babel-preset-expo"],
     plugins: [
       "expo-router/babel",
       ["module:react-native-dotenv", {
