+++ FrontEnd/proxies/LogInServiceProxy.js
@@ -1,5 +1,5 @@
 //import { API_BASE_URL } from '@env';
-const API_BASE_URL = "http://10.34.18.73:5000";
+const API_BASE_URL = "http://192.168.1.67:5000";
 
 const BASE = API_BASE_URL;
