+++ FrontEnd/proxies/ProductosServiceProxy.js
@@ -1,6 +1,6 @@
 //import { API_BASE_URL as BASE } from '@env';
 import { Platform } from 'react-native';
-const BASE = "http://10.34.18.73:5000";
+const BASE = "http://192.168.1.67:5000";
 
 async function _parseBody(response) {
   const text = await response.text().catch(() => '');
