+++ README.md
@@ -1,0 +1,27 @@
+# Manual de instalación
+
+## Repositorio
+
+## Prerrequisitos
+
+A continuación se mostrarán los elementos necesarios para poder ejecutar la aplicación:
+
+### Requerimientos previos:
+
+* Contar con una computadora con Windows, Mac o Linux.
+* Conexión a internet
+
+### Frontend:
+
+* Expo Go
+
+### Backend:
+
+* Node.js
+* npm
+
+### Base de datos:
+
+## Cómo se instala
+
+## Cómo se corre
